Monitoring/requirements.txt
Flask
kazoo
requests
pyjwt
pytz
mysql-connector-python